+++ misskey-api/src/endpoint/admin/roles.rs
@@ -6,3 +6,7 @@ pub mod show;
 pub mod unassign;
 pub mod update;
 pub mod update_default_policies;
+
+#[cfg(feature = "13-7-0")]
+#[cfg_attr(docsrs, doc(cfg(feature = "13-7-0")))]
+pub mod users;