+++ misskey-util/src/builder/admin.rs
@@ -764,6 +764,8 @@ impl<C> AdBuilder<C> {
             priority: Priority::default(),
             #[cfg(feature = "12-81-0")]
             ratio: 1,
+            #[cfg(feature = "13-7-0")]
+            starts_at: DateTime::default(),
             expires_at: DateTime::default(),
             image_url: String::default(),
         };
@@ -839,6 +841,14 @@ impl<C> AdBuilder<C> {
         self
     }
 
+    #[cfg(feature = "13-7-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "13-7-0")))]
+    /// Sets the start date of the ad.
+    pub fn starts_at(&mut self, starts_at: impl Into<DateTime<Utc>>) -> &mut Self {
+        self.request.starts_at = starts_at.into();
+        self
+    }
+
     /// Sets the expiration date of the ad.
     pub fn expires_at(&mut self, expires_at: impl Into<DateTime<Utc>>) -> &mut Self {
         self.request.expires_at = expires_at.into();
@@ -881,6 +891,8 @@ impl<C> AdUpdateBuilder<C> {
     pub fn new(client: C, ad: Ad) -> Self {
         let Ad {
             id,
+            #[cfg(feature = "13-7-0")]
+            starts_at,
             expires_at,
             place,
             priority,
@@ -899,6 +911,8 @@ impl<C> AdUpdateBuilder<C> {
             priority,
             #[cfg(feature = "12-81-0")]
             ratio,
+            #[cfg(feature = "13-7-0")]
+            starts_at,
             expires_at,
             image_url,
         };
@@ -974,6 +988,13 @@ impl<C> AdUpdateBuilder<C> {
         self
     }
 
+    #[cfg(feature = "13-7-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "13-7-0")))]
+    /// Sets the start date of the ad.
+    pub fn starts_at(&mut self, starts_at: impl Into<DateTime<Utc>>) -> &mut Self {
+        self.request.starts_at = starts_at.into();
+        self
+    }
     /// Sets the expiration date of the ad.
     pub fn expires_at(&mut self, expires_at: impl Into<DateTime<Utc>>) -> &mut Self {
         self.request.expires_at = expires_at.into();