+++ misskey-util/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Gallery APIs
 - Reactions APIs
 - Play (Flash) APIs
+- Roles APIs
 
 ### Changed
 ### Deprecated